feat: restructure KafkaJS and ConfluentKafka modules, update imports and documentation
