+++ docgen.json
@@ -1,7 +1,8 @@
 {
   "$schema": "./node_modules/@effect/docgen/schema.json",
   "exclude": [
-    "src/internal/**/*.ts"
+    "src/internal/**/*.ts",
+    "src/*/internal/**/*.ts"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
 }
