+++ examples/kafkajsProducer.ts
@@ -1,13 +1,14 @@
 import { NodeRuntime } from "@effect/platform-node";
 import { Effect } from "effect";
-import { ConfluentKafkaJSInstance, Producer } from "../src";
+import { Producer } from "../src";
+import { ConfluentKafkaJS } from "../src/ConfluentKafka";
 
 const program = Producer.sendScoped({
   topic: "test-topic",
   messages: [{ value: "Hello effect-kafka user!" }],
 }).pipe(Producer.withProducerOptions({ allowAutoTopicCreation: true }));
 
-const KafkaLive = ConfluentKafkaJSInstance.layer({ brokers: ["localhost:19092"] });
+const KafkaLive = ConfluentKafkaJS.layer({ brokers: ["localhost:19092"] });
 const MainLive = Effect.scoped(program).pipe(Effect.provide(KafkaLive));
 
 NodeRuntime.runMain(MainLive);
