+++ examples/streamConsumer.ts
@@ -1,6 +1,7 @@
 import { NodeRuntime } from "@effect/platform-node";
 import { Console, Effect, Stream } from "effect";
-import { Consumer, KafkaJSInstance } from "../src";
+import { Consumer } from "../src";
+import { KafkaJS } from "../src/KafkaJS";
 
 const program = Consumer.serveStream("test-topic").pipe(
   Stream.runForEach(({ topic, partition, ...message }) =>
@@ -14,7 +15,7 @@ const program = Consumer.serveStream("test-topic").pipe(
 );
 
 const ConsumerLive = Consumer.layer({ groupId: "group" });
-const KafkaLive = KafkaJSInstance.layer({ brokers: ["localhost:19092"] });
+const KafkaLive = KafkaJS.layer({ brokers: ["localhost:19092"] });
 const MainLive = program.pipe(Effect.provide(ConsumerLive), Effect.provide(KafkaLive));
 
 NodeRuntime.runMain(MainLive);
