+++ src/ConfluentKafka.ts
@@ -1,0 +1,12 @@
+/**
+ * @since 0.5.0
+ */
+export * as ConfluentKafkaJS from "./ConfluentKafka/ConfluentKafkaJSInstance";
+/**
+ * @since 0.5.0
+ */
+export * as ConfluentRdKafka from "./ConfluentKafka/ConfluentRdKafkaInstance";
+/**
+ * @since 0.5.0
+ */
+export * as Errors from "./ConfluentKafka/ConfluentRdKafkaErrors";
