+++ src/ConfluentKafka/ConfluentKafkaJSInstance.ts
@@ -3,12 +3,12 @@
  */
 import { KafkaJS } from "@confluentinc/kafka-javascript";
 import { Chunk, Config, Effect, Fiber, Layer, Queue, Runtime, Stream } from "effect";
-import * as Consumer from "./Consumer";
-import * as ConsumerRecord from "./ConsumerRecord";
+import * as Consumer from "../Consumer";
+import * as ConsumerRecord from "../ConsumerRecord";
+import * as KafkaInstance from "../KafkaInstance";
+import type * as MessageRouter from "../MessageRouter";
+import * as Producer from "../Producer";
 import * as internal from "./internal/confluentKafkaJSInstance";
-import * as KafkaInstance from "./KafkaInstance";
-import type * as MessageRouter from "./MessageRouter";
-import * as Producer from "./Producer";
 
 /**
  * @since 0.2.0
