+++ src/ConfluentKafka/ConfluentRdKafkaErrors.ts
