+++ src/ConfluentKafka/ConfluentRdKafkaInstance.ts
@@ -3,12 +3,12 @@
  */
 import { GlobalConfig } from "@confluentinc/kafka-javascript";
 import { Array, Chunk, Config, Effect, Fiber, Layer, Queue, Runtime, Stream } from "effect";
-import * as Consumer from "./Consumer";
-import * as ConsumerRecord from "./ConsumerRecord";
+import * as Consumer from "../Consumer";
+import * as ConsumerRecord from "../ConsumerRecord";
+import * as KafkaInstance from "../KafkaInstance";
+import type * as MessageRouter from "../MessageRouter";
+import * as Producer from "../Producer";
 import * as internal from "./internal/confluentRdKafkaInstance";
-import * as KafkaInstance from "./KafkaInstance";
-import type * as MessageRouter from "./MessageRouter";
-import * as Producer from "./Producer";
 
 /**
  * @since 0.4.1
