+++ src/KafkaJS.ts
@@ -1,0 +1,8 @@
+/**
+ * @since 0.5.0
+ */
+export * as KafkaJS from "./KafkaJS/KafkaJSInstance";
+/**
+ * @since 0.5.0
+ */
+export * as Errors from "./KafkaJS/KafkaJSErrors";
