+++ src/KafkaJS/internal/kafkaJSInstance.ts
@@ -12,7 +12,7 @@ import type {
   RecordMetadata,
 } from "kafkajs";
 import { Kafka, logLevel } from "kafkajs";
-import * as Error from "../ConsumerError";
+import * as Error from "../../ConsumerError";
 import { KafkaJSConnectionError, KafkaJSNonRetriableError } from "../KafkaJSErrors";
 
 /** @internal */
